server/symbolic_model/generate.py
# Simple generate placeholder - copies input MIDI to output for demo
import pretty_midi, os, yaml, torch
from model import MusicTransformer

def improve_midi_sync(input_midi, output_midi):
    # For demo purposes, we'll just copy input -> output and run autocorrect if available.
    import shutil
    shutil.copy(input_midi, output_midi)
    # Real implementation: load checkpoint, convert MIDI to tokens, run model, decode tokens back to MIDI.
