requirements.txt
flask
flask-cors
torch>=2.2
torchaudio
pretty_midi
numpy
tqdm
PyYAML
python-dotenv
soundfile

fastapi
uvicorn[standard]
